+++ bin/create.js
@@ -4,7 +4,7 @@ import sade from "sade"
 import { createHandler as handler } from "./handler.js"
 
 sade("create [dir]")
-	.version("v0.4.0")
+	.version("v0.4.1")
 	.describe("quickly create a project template")
 	.example("")
 	.example("new")